+++ packages/cli/.eslintrc.js
@@ -1,7 +1,7 @@
 module.exports = {
   rules: {
     '@typescript-eslint/no-empty-function': 'off',
-    '@typescript-eslint/ban-ts-ignore': 'off',
+    '@typescript-eslint/ban-ts-comment': 'off',
     '@typescript-eslint/explicit-function-return-type': 'off'
   },
   overrides: [