+++ packages/cli/src/lib/webdriver.ts
@@ -6,16 +6,13 @@ import { WebdriverConfigParams } from '../types';
 const startDriver = async (
   config: WebdriverConfigParams
 ): Promise<WebDriver> => {
-  const scriptTimeout = (config.timeout || 20) * 1000.0;
+  const scriptTimeout = config.timeout * 1000.0;
   let builder: Builder;
   /* istanbul ignore else */
   if (config.browser === 'chrome-headless') {
     const service = new chrome.ServiceBuilder(
       config.chromedriverPath || chromedriver.path
-    ).build();
-    // Pinned to selenium-webdriver@4.3.0
-    // https://github.com/SeleniumHQ/selenium/pull/10796/files#diff-6c87d95a2288e92e15a6bb17710c763c01c2290e679beb26220858f3218b6a62L260
-    chrome.setDefaultService(service);
+    );
 
     let options = new chrome.Options().headless();
     if (config.chromeOptions?.length) {
@@ -24,7 +21,10 @@ const startDriver = async (
       }, options);
     }
 
-    builder = new Builder().forBrowser('chrome').setChromeOptions(options);
+    builder = new Builder()
+      .forBrowser('chrome')
+      .setChromeOptions(options)
+      .setChromeService(service);
   } else {
     builder = new Builder().forBrowser(config.browser);
   }