+++ packages/webdriverio/src/test.ts
@@ -13,6 +13,8 @@ import AxeBuilder from '.';
 import { logOrRethrowError } from './utils';
 import { WdioBrowser } from './types';
 import type { AxeResults, Result } from 'axe-core';
+import child_process from 'child_process';
+import { ChildProcessWithoutNullStreams } from 'child_process';
 
 const connectToChromeDriver = (port: number): Promise<void> => {
   let socket: net.Socket;
@@ -50,14 +52,20 @@ describe('@axe-core/webdriverio', () => {
   for (const protocol of ['devtools', 'webdriver'] as const) {
     if (protocol === 'webdriver') {
       port = 9515;
+
+      let chromedriverProcess: ChildProcessWithoutNullStreams;
+
       before(async () => {
-        chromedriver.start([`--port=${port}`]);
+        const path = process.env.CHROMEDRIVER_PATH ?? chromedriver.path;
+        chromedriverProcess = child_process.spawn(path, [`--port=${port}`]);
+        chromedriverProcess.stdout.pipe(process.stdout);
+        chromedriverProcess.stderr.pipe(process.stderr);
         await delay(500);
         await connectToChromeDriver(port);
       });
 
       after(() => {
-        chromedriver.stop();
+        chromedriverProcess.kill();
       });
     }
 
@@ -83,6 +91,10 @@ describe('@axe-core/webdriverio', () => {
       );
       const axeForceLegacy = fs.readFileSync(
         path.join(axeTestFixtures, 'axe-force-legacy.js'),
+        'utf8'
+      );
+      const axeLargePartial = fs.readFileSync(
+        path.join(axeTestFixtures, 'axe-large-partial.js'),
         'utf8'
       );
 
@@ -480,6 +492,20 @@ describe('@axe-core/webdriverio', () => {
 
             assert.notEqual(title, 'Error');
             assert.isNull(error);
+          });
+
+          it('handles large results', async function () {
+            /* this test handles a large amount of partial results a timeout may be required */
+            this.timeout(100_000);
+            await client.url(`${addr}/external/index.html`);
+
+            const results = await new AxeBuilder({
+              client,
+              axeSource: axeSource + axeLargePartial
+            }).analyze();
+
+            assert.lengthOf(results.passes, 1);
+            assert.equal(results.passes[0].id, 'duplicate-id');
           });
 
           it('returns correct results metadata', async () => {